+++ original-version/Measurement.c
@@ -7,12 +7,12 @@ void setMeasurementId(Measurement *measurement, char *id) {
 
 void setDefaultMeasurement(Measurement *measurement) {
 	measurement->adc_value = 0;
-	measurement->_converted_value = 0.0f;
-	measurement->_angular_correction = 1.0f;
-	measurement->_linear_correction = 0.0f;
+	measurement->_converted_value = 0.0;
+	measurement->_angular_correction = 1.0;
+	measurement->_linear_correction = 0.0;
 }
 
-void setMeasurementCorrection(Measurement *measurement, float angular_correction, float linear_correction) {
+void setMeasurementCorrection(Measurement *measurement, double angular_correction, double linear_correction) {
 	measurement->_angular_correction = angular_correction;
 	measurement->_linear_correction = linear_correction;
 }
@@ -21,7 +21,7 @@ void _convertMeasurement(Measurement *measurement) {
 	measurement->_converted_value = measurement->adc_value * measurement->_angular_correction + measurement->_linear_correction; 
 }
 
-float getMeasurementValue(Measurement *measurement) {
+double getMeasurementValue(Measurement *measurement) {
 	_convertMeasurement(measurement);
 	return measurement->_converted_value;
 }
@@ -29,6 +29,6 @@ float getMeasurementValue(Measurement *measurement) {
 void printMeasurement(Measurement *measurement) {
 	printf("Measurement: %s\t", measurement->id);
 	printf("ADC Value: %d\t", measurement->adc_value);
-	printf("Converted Value: %.3f\n", measurement->_converted_value);
+	printf("Converted Value: %.3lf\n", measurement->_converted_value);
 	printf("\n");
 }
